:recycle: refactor: layout component
