+++ packages/ui/src/components/Layout/components/Main/Main.test.tsx
@@ -1,0 +1,9 @@
+import { render } from "@testing-library/react";
+import { composeStories } from "@storybook/testing-react";
+import * as stories from "./Main.stories";
+
+const { Default } = composeStories(stories);
+test("renders", () => {
+  render(<Default />);
+  expect(1).toBe(1);
+});
