+++ packages/ui/src/components/Layout/components/MenuItems/SecondaryListItems.test.tsx
@@ -1,0 +1,9 @@
+import { render } from "@testing-library/react";
+import { composeStories } from "@storybook/testing-react";
+import * as stories from "./SecondaryListItems.stories";
+
+const { Default } = composeStories(stories);
+test("renders", () => {
+  render(<Default />);
+  expect(1).toBe(1);
+});
