+++ packages/ui/src/components/Layout/components/MenuItems/index.ts
@@ -1,0 +1,1 @@
+export * from "./MenuItems";
