+++ packages/ui/src/components/Layout/components/SideBar/SideBar.test.tsx
@@ -1,6 +1,6 @@
 import { render } from "@testing-library/react";
 import { composeStories } from "@storybook/testing-react";
-import * as stories from "./Layout.stories";
+import * as stories from "./SideBar.stories";
 
 const { Default } = composeStories(stories);
 test("renders", () => {
