+++ packages/ui/src/components/Layout/components/SideBar/SideBar.tsx
@@ -1,0 +1,61 @@
+import { Divider, IconButton, List, styled, Toolbar } from "@mui/material";
+import ChevronLeftIcon from "@mui/icons-material/ChevronLeft";
+import MuiDrawer from "@mui/material/Drawer";
+import { Dispatch } from "react";
+import { MainListItems, SecondaryListItems } from "../MenuItems";
+
+const drawerWidth = 240;
+
+const Drawer = styled(MuiDrawer, {
+  shouldForwardProp: (prop) => prop !== "open",
+})(({ theme, open }) => ({
+  "& .MuiDrawer-paper": {
+    position: "relative",
+    whiteSpace: "nowrap",
+    width: drawerWidth,
+    transition: theme.transitions.create("width", {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.enteringScreen,
+    }),
+    boxSizing: "border-box",
+    ...(!open && {
+      overflowX: "hidden",
+      transition: theme.transitions.create("width", {
+        easing: theme.transitions.easing.sharp,
+        duration: theme.transitions.duration.leavingScreen,
+      }),
+      width: theme.spacing(7),
+      [theme.breakpoints.up("sm")]: {
+        width: theme.spacing(9),
+      },
+    }),
+  },
+}));
+
+export type SideBarProps = {
+  open: boolean;
+  toggleOpen: () => void;
+};
+
+export const SideBar = ({ open, toggleOpen }: SideBarProps) => (
+    <Drawer variant="permanent" open={open}>
+      <Toolbar
+        sx={{
+          display: "flex",
+          alignItems: "center",
+          justifyContent: "flex-end",
+          px: [1],
+        }}
+      >
+        <IconButton onClick={toggleOpen}>
+          <ChevronLeftIcon />
+        </IconButton>
+      </Toolbar>
+      <Divider />
+      <List component="nav">
+        <MainListItems />
+        <Divider sx={{ my: 1 }} />
+        <SecondaryListItems />
+      </List>
+    </Drawer>
+  );
