+++ packages/ui/src/components/Layout/components/SideBar/index.ts
@@ -1,0 +1,1 @@
+export * from "./SideBar";
